# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Permy \u2014 Building Permit & Construction Intelligence API",
    "description": "Normalized permits, contractor activity, property timelines, and ZIP development signals \u2014 ranked, sourced, and ready for your app. Give Permy an address, city, ZIP, contractor, or trade and get back clean, machine-readable municipal intelligence with source links and confidence scores.",
    "termsOfService": "https://docs.permy.dev/terms",
    "contact": {
      "name": "Permy",
      "url": "https://permy.dev/",
      "email": "hi@permy.dev"
    },
    "license": {
      "name": "Apache-2.0 (code); data per upstream license",
      "url": "https://docs.permy.dev/legal"
    },
    "version": "0.1.0",
    "x-logo": {
      "url": "https://permy.dev/assets/brand-mark.png"
    }
  },
  "servers": [
    {
      "url": "https://permy.p.rapidapi.com",
      "description": "RapidAPI production"
    },
    {
      "url": "https://api.permy.dev",
      "description": "Direct site"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local dev"
    }
  ],
  "paths": {
...
    "/v1/contractors/search": {
      "get": {
        "tags": [
          "contractors & markets"
        ],
        "summary": "Find contractors by name, license, city, trade, activity",
        "operationId": "search_contractors_v1_contractors_search_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "license",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "License"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "trade",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trade"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractorsSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/contractors/{contractor_id}/activity": {
      "get": {
        "tags": [
          "contractors & markets"
        ],
        "summary": "Permit count, trade mix, active cities, value bands, momentum, source confidence",
        "operationId": "contractor_activity_v1_contractors__contractor_id__activity_get",
        "parameters": [
          {
            "name": "contractor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contractor Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractorActivity"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/markets/{zip}/development-score": {
      "get": {
        "tags": [
          "contractors & markets"
        ],
        "summary": "ZIP-level development momentum: volume, value, trade mix, MoM delta, hotspot 0\u2013100",
        "operationId": "market_score_v1_markets__zip__development_score_get",
        "parameters": [
          {
            "name": "zip",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Zip"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketScore"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/v1/alerts": {
      "post": {
        "tags": [
          "alerts & webhooks"
        ],
        "summary": "Create a saved search with optional webhook delivery",
        "operationId": "create_alert_v1_alerts_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alert"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "alerts & webhooks"
        ],
        "summary": "List your saved searches",
        "operationId": "list_alerts_v1_alerts_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Alert"
                  },
                  "title": "Response List Alerts V1 Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/alerts/{alert_id}": {
      "delete": {
        "tags": [
          "alerts & webhooks"
        ],
        "summary": "Delete a saved search",
        "operationId": "delete_alert_v1_alerts__alert_id__delete",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/webhooks/test": {
      "post": {
        "tags": [
          "alerts & webhooks"
        ],
        "summary": "Send a signed HMAC test payload to your webhook URL",
        "operationId": "webhook_test_v1_webhooks_test_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Alert": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "persona": {
            "type": "string",
            "enum": [
              "roofer",
              "solar",
              "hvac",
              "investor",
              "supplier",
              "insurer",
              "general"
            ],
            "title": "Persona"
          },
          "query": {
            "additionalProperties": true,
            "type": "object",
            "title": "Query"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "last_fired_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Fired At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "persona",
          "query",
          "created_at"
        ],
        "title": "Alert"
      },
      "AlertCreate": {
        "properties": {
          "persona": {
            "type": "string",
            "enum": [
              "roofer",
              "solar",
              "hvac",
              "investor",
              "supplier",
              "insurer",
              "general"
            ],
            "title": "Persona",
            "default": "general"
          },
          "query": {
            "additionalProperties": true,
            "type": "object",
            "title": "Query",
            "description": "Same params shape as /v1/permits/search"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "webhook_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Secret"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "AlertCreate"
      },
      "Contractor": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "canonical_uid": {
            "type": "string",
            "title": "Canonical Uid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number"
          },
          "license_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License State"
          },
          "license_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Status"
          },
          "trade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "zip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "canonical_uid",
          "name"
        ],
        "title": "Contractor"
      },
      "ContractorActivity": {
        "properties": {
          "contractor": {
            "$ref": "#/components/schemas/Contractor"
          },
          "permit_count": {
            "type": "integer",
            "title": "Permit Count",
            "default": 0
          },
          "total_valuation_usd": {
            "type": "number",
            "title": "Total Valuation Usd",
            "default": 0.0
          },
          "trade_mix": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Trade Mix"
          },
          "active_cities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Cities"
          },
          "value_band": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "<50k",
                  "50k-500k",
                  "500k+"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Band"
          },
          "momentum": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Momentum",
            "description": "Recent activity intensity 0\u20131",
            "default": 0.0
          },
          "first_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen At"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "contractor"
        ],
        "title": "ContractorActivity"
      },
...
      "ContractorsSearchResponse": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "contractors": {
            "items": {
              "$ref": "#/components/schemas/Contractor"
            },
            "type": "array",
            "title": "Contractors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "contractors"
        ],
        "title": "ContractorsSearchResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketScore": {
        "properties": {
          "zip": {
            "type": "string",
            "title": "Zip"
          },
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "permit_count_30d": {
            "type": "integer",
            "title": "Permit Count 30D"
          },
          "permit_count_90d": {
            "type": "integer",
            "title": "Permit Count 90D"
          },
          "total_value_30d": {
            "type": "number",
            "title": "Total Value 30D"
          },
          "total_value_90d": {
            "type": "number",
            "title": "Total Value 90D"
          },
          "trade_mix": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Trade Mix"
          },
          "mom_delta_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mom Delta Pct",
            "description": "Month-over-month permit volume %"
          },
          "top_contractors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Top Contractors"
          },
          "hotspot_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Hotspot Score",
            "description": "ZIP development momentum 0\u2013100"
          },
          "narrative": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative",
            "description": "One-line human summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "zip",
          "as_of_date",
          "permit_count_30d",
          "permit_count_90d",
          "total_value_30d",
          "total_value_90d",
          "hotspot_score"
        ],
        "title": "MarketScore"
      },
...
      "WebhookTestRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url"
        ],
        "title": "WebhookTestRequest"
      },
      "WebhookTestResponse": {
        "properties": {
          "delivered": {
            "type": "boolean",
            "title": "Delivered"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "delivered"
        ],
        "title": "WebhookTestResponse"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Key",
        "description": "Your RapidAPI app key (sent automatically by the gateway). Direct-site callers may use X-API-Key or Authorization: Bearer instead."
      }
    }
  },
...
}